refactor: :recycle: create lib for github
